+++ protocols/gossipsub/CHANGELOG.md
@@ -1,4 +1,7 @@
 ## 0.50.0
+- Prevent mesh exceeding mesh_n_high.
+  See [PR 6184](https://github.com/libp2p/rust-libp2p/pull/6184)
+
 - Fix underflow when shuffling peers after prunning.
   See [PR 6183](https://github.com/libp2p/rust-libp2p/pull/6183)
 
@@ -16,12 +19,18 @@
 
 - Fix incorrect default values in ConfigBuilder
   See [PR 6113](https://github.com/libp2p/rust-libp2p/pull/6113)
-  
+
 - Remove duplicated config `set_topic_max_transmit_size` method, prefer `max_transmit_size_for_topic`.
   See [PR 6173](https://github.com/libp2p/rust-libp2p/pull/6173).
 
 - Switch the internal `async-channel` used to dispatch messages from `NetworkBehaviour` to the `ConnectionHandler`
   with an internal priority queue. See [PR 6175](https://github.com/libp2p/rust-libp2p/pull/6175)
+
+- gossipsub: do early return in for an empty input
+  See [PR 6208](https://github.com/libp2p/rust-libp2p/pull/6208).
+
+- Refactor gossipsub with in-place negative-score peer removal.
+  See [PR 6209](https://github.com/libp2p/rust-libp2p/pull/6209).
 
 ## 0.49.2